feita os códigos pageController.js e pageModel.js
